+++ lib/src/socket.dart
@@ -129,10 +129,6 @@ class Socket extends EventEmitter {
     emitWithAck(event, data);
   }
 
-  void emitWithBinary(String event, [data]) {
-    emitWithAck(event, data, binary: true);
-  }
-
   ///
   /// Emits to this client.
   ///
@@ -153,7 +149,7 @@ class Socket extends EventEmitter {
       }
 
       var packet = {
-        'type': binary ? BINARY_EVENT : EVENT,
+        'type': EVENT,
         'data': sendData,
         'options': {'compress': flags?.isNotEmpty == true && flags['compress']}
       };
@@ -192,13 +188,13 @@ class Socket extends EventEmitter {
     _logger.fine('transport is open - connecting');
 
     // write connect packet if necessary
-    if ('/' != nsp) {
-      if (query?.isNotEmpty == true) {
-        packet({'type': CONNECT, 'query': query});
-      } else {
-        packet({'type': CONNECT});
-      }
+    // if ('/' != nsp) {
+    if (query?.isNotEmpty == true) {
+      packet({'type': CONNECT, 'query': query});
+    } else {
+      packet({'type': CONNECT});
     }
+    // }
   }
 
   ///
@@ -248,7 +244,7 @@ class Socket extends EventEmitter {
         ondisconnect();
         break;
 
-      case ERROR:
+      case CONNECT_ERROR:
         emit('error', packet['data']);
         break;
     }